+++ GitUI/GitUI.csproj
@@ -50,7 +50,7 @@
     <!-- Resources -->
     <EmbeddedResource Update="Properties\Resources.resx" Generator="ResXFileCodeGenerator" LastGenOutput="Resources.Designer.cs" />
     <Compile Update="Properties\Resources.Designer.cs" AutoGen="True" DependentUpon="Resources.resx" DesignTime="True" />
-    <Compile Update="UserControls\GPGSecretKeysComboboxControl.cs">
+    <Compile Update="UserControls\GPGKeys\GPGSecretKeysCombobox.cs">
       <SubType>UserControl</SubType>
     </Compile>
