--- GitUI/Properties/DataSources/GitUI.UserControls.GPGSecretKeysComboboxControl+GpgKeyDisplayInfo.datasource
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<!--
-    This file is automatically generated by Visual Studio .Net. It is 
-    used to store generic object data source configuration information.  
-    Renaming the file extension or editing the content of this file may   
-    cause the file to be unrecognizable by the program.
--->
-<GenericObjectDataSource DisplayName="GPGSecretKeysComboboxControl+GpgKeyDisplayInfo" Version="1.0" xmlns="urn:schemas-microsoft-com:xml-msdatasource">
-   <TypeInfo>GitUI.UserControls.GPGSecretKeysComboboxControl+GpgKeyDisplayInfo, GitUI, Version=3.3.0.0, Culture=neutral, PublicKeyToken=null</TypeInfo>
-</GenericObjectDataSource>
